create new eventhubmetricsprovider, use in scalemonitor, update to prerease webjobs sdk with target scaler interface
